Improves the robustness of the system start by handling invalid config files gracefully
If one of the config files (application.conf, instance.conf or component-*.conf)
contained a syntax error, the whole startup process crashed. We now simply report the
error and skip the file, so that the system at least triest to start. We also directly
use a logger and not Exceptions, because this might not be properly initialized at
this early stage...