Mark failed Promise as non successful
A failed promise was marked as successful. This could happen if the promise was marked as successful with a value and in one of the success handlers an exception is thrown. In this case the promise is marked as successful and failed at the same time. If a fail handler is attached after this, it would be executed as if the promise was successful. With this change, the handler in this case would be executed as failed.