Transforming also checks superclass
In some cases there might be a need to transform two classes which are not directly bound by a transformer e.g. if the source class is a more specific implementation. In this case it was needed to write multiple transformers in order to support this case. With these changes not only the source class itself is checked for transforming, but also its superclass. If this is not desired, the transformer can indicate this by setting a flag.