Additional transformation sources
Sometimes child classes of a class should be transformed as well as their parent class. In order to avoid writing multiple transformers for each case, transformers now support adding multiple classes additionally as source classes for transforming them into the target class. The only restriction is that the additional classes have to extend the parent class.