Allow transformers for child classes
Sometimes there is only a transformer for a parent class to another class. In order to avoid writing multiple transformers for these child classes, transformers now walk the inheritance path down until the visited class can be transformed into the target class or the class to check does not implement the Transformable interface in order to avoid unnecessary transformation attempts.