Permits the defer the language detection.
Many tasks don't need the language at all (e.g. the web-server serving
an image). On the other hand, determining the current language is quite
expensive (checking a user-session / parsing the accept language header).

Therefore we permit to install a handler which is only evaluated if
and when required. We also install the default language before doing so,
as otherwise endless recursions might occur.