Fixes Bug in which sirius could not stop on Sirius::stop
The problem was, there were two Default-Executor created.

Sirius starts Default-Executor before Tasks::Started and adds it to the Tasks.executers map. In Tasks::Started the map is cleared. When another Default-Executor task is started, another Default-Executor is created. Only one of those two Default-Executors is in the Tasks.executers map and gets shut down.
The other one is not shut down, because it is not in the Task.executers-Map and hinders the program to terminate.