Returns empty string instead of trowing exception when key is empty
As looking up an empty string in the translations map won't ever result in a found translation and instead an exception is thrown that is of no use we instead pass empty string back.
This is useful for example in the i18n macro of sirius-web as we very often use an empty string as default when no value is passed for an optional parameter.