Cleanup of the Babelfish.
Previously we stored additional metadata like "file" and "accessed
to be able to output a set of valid properties files again. We now
use an external translation facility which is highly encouraged.

Additionally some methods were broken down to encourage
inlining by the JIT and also the initial load of all properties is
done in one lock to prevent a log of garbage during startup.