Adds a flag which lets all docker containers continue to run.
This GREATLY increases the developer productivity as it minimizes
the restart time of the debugger in case of a chance which cannot be
performed via hot code replacement.

This _can_ also be specified when executing tests although this should
be handled with care and not be enabled on a CI system.