Makes all getXXX method behave the same way when null is passed in as default.
We now accept (and then return) null as valid default value. We therefore removed
the Nonnull constraints and guarantees as the output depends on the input value.

Every method except one already handled this case gracefully and this reduced the
total complexity and number of methods in this already complex class.