Provides a more robust way to lazily initializing the language.
We support a pattern to only initialize the language und demand. This
is used by the web server, so that we do not have to perform a full
user lookup and validation for static assets and the like.

However, many of these lazy language installers will only transfer
the user language if it has not been set before (using setLangIfEmpty).

Therefore we now no longer pre-initialize the lang with a default value
but rather set the installer to null to break a potential endless recursion.

If the installer doesn't populate the language, we still use the default
language then...