Provides a cluster capable way of removing a bunch of entries from a cache.
We previously used Cache.removeIf which ran a local predicate over all entries
and then only distributed the resulting deletes. However, this leads to a cache coherence
bug and might result in stale cache entries. This happens if the node being scanned contains
less or other cache entries than its peer nodes. These other keys and their entries are then
never scanned and will therefore be missed.

We now take another approach: We register a handler (or several) on each node. This
handler has a name and a predicate which scans and removes all cache entries if
the match a given string input.

Additionally a second method "removeAll" is added where we can specify the name of the
delete handler to use and the test input into the predicate. These two values can easily be
distributed in a cluster and then each cache can locally scan and remove all matching entries.

Cache.removeIf has been marked deprecated and will be removed soon!