Provides a simple facility to execute daily cleanup and maintenance tasks.
Most systems need to perform regular cleanup activities. Previously these
have been scheduled as "EveryDay" tasks, where the user had to chose
an hour of execution wisely. We now collect all these tasks and execute
them in a defined time window (10pm - 5am by default). If the system is
unable to perform all tasks within this window, execution is stopped and
continued in the next night.

All tasks are shuffled on each run to prevent that a single task steals
all the time.

Note however, that once a single task is started, it is always completely
executed and it is the duty of the programmer to limit the execution time
of each task itself - if necessary.