Fetches the current language as late as possible.
Many formatters are used in system code (e.g. when initializing
a Database in sirius-db). As long as we're just juggling strings
or only rely on setDirect, there is no need to determine the current
language (as this might invoke a complex piece of code which
then resolves the current user etc.)