BREAKING CHANGE: Changes the way @Register detects classes of interest.
Previously @Register took only the list of interfaces and registered the
newly instantiated part for all of these. This lead to a pretty common
pitfall if the target to register for was the class itself, another class
or an interface implemented by a common parent class.

Therefore we now provide the marker annotation @AutoRegister which
can be placed on all interfaces or classes which should end up in
the MutableGlobalContext.registerPart call.

To register for an interface which is implemented by this part, simply
add @Register to the part and @AutoRegister to the interface
(same as before).

To also register to the part class itself, simply add @AutoRegister next
to @Register on the part itself.

To register for an interface implemented by a parent class (e.g.
subclasses of BizController in sirius-biz), simply use @Register
on the controller as the Controller interface (will) wear @AutoRegister).

To register for a class (e.g. BackgroundLoop) also simply use
@Reigster on the loop part, as BackgroundLoop will also wear an
@AutoRegister annotation.