Replaces log4j as logging framework.
As we're now only log to the console (and let Docker manage
everything else), we have no need for a sophisticated framework
like log4j. Therefore we simply use the inner workings of the
JDK/JRE with als little code as possible.

There are no real changes concerning the framework as we
abstracted logging in our "Log" class, where only the enum
defining the log levels changes.