Removes unnecessary Guava calls.
We used this heavily in the pre Java8 world as
the diamond operator wasn't there yet and one
was forced to re-type all generics.

This is now gone so that we can stick with plain
Java calls. We also replaced calls to Hashing and
ByteStreams by our own (non beta, non deprecated
implementations).