Prevents overuse of Redis via coherent caches
The change in the reverted commit was meant to flush an element from the cache on all nodes when it was newly loaded on a node. This actually resulted in a loop of Redis messages when an cached element was needed on multiple nodes and also made the caches for such elements quite ineffective.

Reverts changes of 92d5d86493a5677b510695efa6776e7cdf81bb1b

Fixes: SIRI-234