BREAKING: No longer provides access to outputstream for get requests
Previously this would work, as the request was transformed into POST from GET if you called getOutput on a GET connection. As we have an explicit method to transform a connection into a POST connection, we excpect this method to be called instead of relying on this auto transform (which is marked as backwards compatibility in the java API).