BREAKING CHANGE: Properly handles rounding in Amount.
Previously, calling round didn't change the scale of the internal
number and was thus lost for all subsequent calls.

This meant that formatting this number didn't yield the expected result
as e.g. putting Amount(1.33333).round(3) into an XML response did
yield "1,33" or if getAmount() was used 1.33000 but not the expected
"1.333". This has now been fixed.

Also we no use the machine format when writing XML instead of the user
readable (this is the breaking change).