Updates to the latest Guava version to mitigate a security risk.
CVE-2020-8908
moderate severity
Vulnerable versions: <= 29.0
Patched version: 30.0-jre

A temp directory creation vulnerability exist in Guava versions prior
to 30.0 allowing an attacker with access to the machine to potentially
access data in a temporary directory created by the Guava
com.google.common.io.Files.createTempDir(). The permissions granted to
 the directory created default to the standard unix-like /tmp ones,
leaving the files open. We recommend updating Guava to version 30.0 or
later, or update to Java 7 or later, or to explicitly change the
permissions after the creation of the directory if neither are possible.