Prevents that one accidentally adds a Value or Optional to a context.
In the past, this was a bug in 100% of all cases, as consumers of the
Context normally expect plain types and not ones wrapped in a Value
or Optional. Also there is no actual benefit in doing so - however, there
is no real support in the Java language to prevent this.

Therefore we add these checks to at least "fail early and notably".