Overrides existing headers in Outcalls
Before we switched to HttpClient, headers were always overridden when a
header with the same name already existed. Now a new header is added
every time setRequestProperty is called. This materialized as a bug in
the JSONCall if instead of JSON, parameters were POSTed. The JSON#to
method would set the Content-Type header to application/json and
Outcall#postData would add a second Content-Type header with
application/x-www-form-urlencoded, which led to an invalid request.
This commit reverts to the previous behavior.

Fixes: SE-11264