Allows to define a custom processor for XML calls.
We may require custom handling of the input stream before we can consider it valid XML. This was possibly by manually handling the input stream of the corresponding Outcall, but this lead to bypassing any logging.
Note, that input is FINE logged before processing to make sure that we log input as is and to even log in cases where the processor throws an exception.

By default, the input stream is simply passed through without change.

Fixes: SE-12549