Moves creation of the outcall to the constructor.
This is necessary as otherwise, there is no way for extensions of this class to create the outcall without also creating an XMLCall instance.
Extending classes can simply call the super-constructor with the target URI and content type now.

This also allows the Outcall to be final and gets rid of a useless default constructor (as the default constructor never set the outcall and the class doesn't allow creation of the outcall later).

This is technically breaking, but as the default constructor never worked properly, it is highly unlikely anyone ever used it instead of the static `to` methods that created the XMLCall objects.

Fixes: SE-12549