Unifies retrieval of arrays and objects via jackson
The different methods now all provide a try-variant that return an empty optional when no fitting array/object node could be retrieved.
The functionality of the methods without the try-prefix is changed so they fall back to an empty array/node when the try-variant didn't find a result.
This replaces the old usage of withArray/withObject that actually added the array/object to the parent the retrieval was called on.

Fixes: SIRI-736