Provides a central string normalization and cleanup method.
We have multiple "normalizeString" method sprinkled accross our
codebase - and we never know what "normalize" actually means.

Therefore, we provide a central method with explicit cleanup
parameters. Note that the enum sets defining the operations
should probably be defined as constants so that we always
apply the same op for the same strings.