Expands the sub-context helpers of the CallContext
This allows to check whether a sub-context of a specific type is currently attached to the call context and allows to directly retrieve it without the safety checks and creation of new instance if none is attached.

It also provides a way to cleanly remove a sub-context from the call context, without calling set with null.

Fixes: OX-10601