Provide helper for quoting spaces in URLs
We known that spaces in URLs are prohibited but in some cases they are provided as such in XML-files.
This is necessary as the URI constructor that handles a full URL string does not do this automatically.
Before the deprecation of the URL constructors we would use a workaround where we first create a URL and then pass all components of the URL in the specific URI constructor which would escape the path and query accordingly.
With the deprecation that logic was switched back to the URI constructor and the functionality broke.

With this helper we now can explicitly force this behavior where necessary.

Fixes: OX-11739