Adds an explicit constant to better document behavior
Outcall.modifyClient(String) accepts a string to cache a
newly constructed HttpClient. null is used to signal that
the client must not be cached. To better document this,
an explicit constant was added so that

call.modifyClient(Outcall.CLIENT_SELECTOR_NO_CACHE)
sort of speaks for itself (in contrast to call.modifyClient(null)